requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
discord.py==1.7.3
idna==3.3
multidict==6.0.2
pygame==2.1.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.2
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
